apps/backend/requirements.txt
# FastAPI 核心
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
pyotp>=2.9.0  # TOTP 双因素认证

# 配置
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0

# Redis
redis>=4.5.0

# LLM
cryptography>=41.0.0
httpx>=0.24.0

# HTML 解析（导入器依赖）
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 开发工具
pytest>=7.0.0
